Submitted by:  Robert Haschart
Minor change to the StreamReader so that the library can read from a pipe.  Previously the code used inputstream.available() to determine when it was done reading from a file, but when reading from a pipe,  available() == false  can indicate that the pipe doesn't yet have data available, but it will have data in the future.   Now it calls mark() on the stream,  tries to read a byte, which will return -1 at EOF, and block until a byte is available, and then calls reset on the stream to restore that byte to the stream.