Submitted by:  Robert Haschart
Allow specification of encodings other than those that are officially supported in Marc records, relies on the java character encoding, and only supports those additional encodings supported by the installed version of java.   Additionally look for and delete extraneous characters that occur between Marc records in a file of Marc records.  (for instance if CR LF is inserted between records.)