Submitted by:  Robert Haschart
If a Unicode code point doesn't have a corresponding Ansel representation, first try performing a Unicode decompose normalization, and try to encode the result as Ansel.  If the decomposed form doesn't help, then encode the codepoint directly as &#xDDDD;    (where D is a hexadecimal digit  0-9A-F)

This latest change merely ensures that if the Unicode code point to be encoded as &#xDDDD; is less than or equal to  0xFFF, the encoding will still be well formed.