We check for field indicators, need the same check here
Error tossed by MarcXmlHandler was generic "unable to parse input"

So we'll check for null as w/ field, except that lets not throw out the entire record batch just b/c
of a bad subfield.  Coerce the code to a space when missing/null.
Imperfect, but better than inspecific crashing on a common enough
problem.