Update MarcJsonWriter.java
Added escaping for more components of a marc record.
Closed the bounds check for characters that need escaping to include 0x1f.
Added symmetric checks for valid tags and subfields that are present in MarcJsonReader.
Added missing assignment of the output stream in the last constructor.
Replaced blocks for indenting with calling the existing indent method.
Replaced getting the Unicode bytes by name with getting the bytes using the standard character set for Unicode.